Current time feature and Clubbing all open application/url in one.
New option to get the current time of the day added.
Combined all different open <link> or open <application> queries into one single query.
Checks if given target is an application present in AppPath dictionary in ExternalPath.py and opens it if found.
Checks if given target is a web url present in WebPath dictionary in ExternalPath.py and opens it if found.
Otherwise informs user that it could not resolve the target.